Final touch of the backend API - Writing documentation and Readme.md - Finalising unittest - Various changes done to accomodate Frontend integration

+++ ToDoList.Application/ToDoTaskLists/Commands/DeleteToDoTaskList/DeleteToDoTaskListHandler.cs
@@ -12,7 +12,6 @@ namespace ToDoList.Application.ToDoTaskLists.Commands.DeleteToDoTaskList;
 
 public class DeleteToDoTaskListHandler : IRequestHandler<DeleteToDoTaskList, ToDoTaskListDto>
 {
-    private readonly IToDoTaskRepository _toDoTaskRepository;
     private readonly IToDoTaskListRepository _toDoTaskListRepository;
     private readonly IMapper _mapper;
 
@@ -24,7 +23,11 @@ public class DeleteToDoTaskListHandler : IRequestHandler<DeleteToDoTaskList, ToD
 
     public async Task<ToDoTaskListDto> Handle(DeleteToDoTaskList request, CancellationToken cancellationToken)
     {
-        var taskList = await _toDoTaskListRepository.GetTaskListByIdAsync(request.ToDoTaskListDto.Id);
+        var taskList = await _toDoTaskListRepository.GetTaskListByIdAsync(request.Id);
+        if (taskList == null)
+        {
+            throw new KeyNotFoundException($"Task list with Id '{request.Id}' not found.");
+        }
         await _toDoTaskListRepository.DeleteTaskListAsync(taskList);
         return _mapper.Map<ToDoTaskListDto>(taskList);
     }
